requirements.txt
# Academy Program Backend - Production Dependencies
# Generated for Render deployment

# Web Framework
fastapi==0.128.0
uvicorn==0.40.0
starlette==0.50.0

# Data Validation
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.3.0

# PostgreSQL (async)
asyncpg==0.31.0
SQLAlchemy==2.0.45
greenlet==3.3.0

# MongoDB (async)
motor==3.7.1
pymongo==4.15.5
dnspython==2.8.0

# Redis (async)
redis==7.1.0

# Authentication
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.21

# Environment
python-dotenv==1.2.1

# HTTP Client (for testing)
httpx==0.28.1

# Cryptography dependencies
cryptography==46.0.3
